% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetools.R
\name{are.settings.and.env.correct}
\alias{are.settings.and.env.correct}
\title{chech the settings and environment exists}
\usage{
are.settings.and.env.correct()
}
\value{
TRUE - fine. False - not fine
}
\description{
chech the settings and environment exists
}
